DOKY-276 Migrate search functionality from Azure to Algolia
- Replaced Azure Search with Algolia for search and indexing services.
- Removed Azure-specific configurations, dependencies, and classes.
- Added support for Algolia-based indexing and search services.
- Updated `DocumentResultData` and `DocumentIndexData` to align with Algolia's schema.
- Adjusted application properties for Algolia configurations.
- Updated tests to reflect changes in indexing and search implementations.

+++ src/main/kotlin/org/hkurh/doky/documents/api/DocumentSearchResponse.kt
@@ -27,5 +27,5 @@ package org.hkurh.doky.documents.api
  */
 data class DocumentSearchResponse(
     val documents: List<DocumentResponse>,
-    val totalCount: Long
+    val totalCount: Int
 )
